requirements.txt
Django>=4.2.0
django-rest-framework>=3.14.0
django-cors-headers>=4.1.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
Pillow>=9.5.0